SentimentAnalysis/nlp/vader_sentiment.py
from nltk.sentiment.vader import SentimentIntensityAnalyzer
import nltk

# Download once
nltk.download('vader_lexicon', quiet=True)

analyzer = SentimentIntensityAnalyzer()

def analyze_sentiment(text):
    scores = analyzer.polarity_scores(text)
    compound = scores['compound']
    
    if compound >= 0.05:
        sentiment = "Positive"
    elif compound <= -0.05:
        sentiment = "Negative"
    else:
        sentiment = "Neutral"
    
    return sentiment, compound
