SentimentAnalysis/requirements.txt
numpy==1.23.5
pandas==1.5.3
opencv-python==4.7.0.72
tensorflow==2.13.0
keras==2.13.1
flask==3.1.1
streamlit
Pillow
vaderSentiment